refactor: Remove check-dependencies task for now
We always install latest dependencies so it's not super useful to check vulnerabilities I guess, since we can't upgrade to safer versions.

Also safety depends on a pre-release of dparse, which is causing resolution issues.

And I'm not really a fan of safety anymore. I'm planning to switch to something else.