refactor: Rewrite make script in Python
I kept having issues with options parsing in Bash.

We have to use double-quotes around `${var[@]}` so that each item is correctly passed as a whole argument, even if it contains space.

This works well when we parse a single option into our array of options. But with more than one option, it expands with new lines between each argument...

I grew tired of these quirks so just rewrote the thing in Python.